Refactor: 업적 관련 enum을 type 디렉토리로 이동
