+++ src/types/Game.ts
@@ -1,4 +1,4 @@
-import { AchievementDescription, AchievementName } from '../utils/achievements';
+/* eslint-disable no-unused-vars */
 import { UserInfoType } from './User';
 
 export type MatchType = {
@@ -11,6 +11,22 @@ export type MatchType = {
   winner: string | null,
   loser: string | null,
 };
+
+export enum AchievementName {
+  FIRST_GAME = 'FIRST_GAME',
+  FIRST_WIN = 'FIRST_WIN',
+  FIRST_LOSE = 'FIRST_LOSE',
+  FIRST_FRIEND = 'FIRST_FRIEND',
+  FIRST_BLOCK = 'FIRST_BLOCK',
+}
+
+export enum AchievementDescription {
+  FIRST_GAME = 'You have successfully played your first game.',
+  FIRST_WIN = 'Congratulations on your first win.',
+  FIRST_LOSE = 'You have achieved your first lose.',
+  FIRST_FRIEND = 'You made your first friend.',
+  FIRST_BLOCK = 'You made your first hater.',
+}
 
 export type AchievementType = {
   name: AchievementName,
